pciutils: Add make-native dependency
A change in behaviour in make between 4.2.1 and 4.3 on how whitespace
and appends are handled[1] causes changes in lib/libpci.pc and leads
to non-reproducible builds.

Add a dependency on make-native to resovle this as a least invasive
and least worse solution for now.

[1] https://git.savannah.gnu.org/cgit/make.git/commit/?id=b90fabc8d6f34fb37d428dc0fb1b8b1951a9fbed

(From OE-Core rev: b7c366036d6e93854d976678dbe770039de68725)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>