autotools: Drop relative configure path workaround
Calling configure with a relative path to the build directory was a nice
workaround to remove hardcoded paths from files. We have proper prefix
remapping code now and these relative paths escape that modern code.
Drop the obsolete approach and allow the prefix remapping to handle things.
This ensures the contents of debug source directories is correct as the build
directory from the tarball name may not match the on target debug directory.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>