rng-tools: Remove disabling DefaultDependencies
Delete dependency on dev-hwrng.device and systemd-udev-settle services
as they can be troublesome because difference devices may have different
devices to detect resulting in delayed boot. default dependencies would
add some delay that the race condition for creating the device by kernel
module and daemon using it is not seen often.

do not run if included in a container image

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>