ref-manual: complementary package installation recommends
Add a note to the COMPLEMENTARY_GLOB variable glossary entry mentioning
that recommends aren't installed (since this seems to be the only place
that complementary package installation is really discussed). Also turn
the paragraph about globbing into a regular paragraph as I don't think
it should be called out in a note block.

Signed-off-by: Paul Eggleton <paul.eggleton@microsoft.com>