oeqa/selftest/runtime_test: correctly inherit testimage.bbclass
Since oe-core 506346, testexport.bbclass pulls in testimage.bbclass via
IMAGE_CLASSES.  This works if testexport is inherited with an 'inherit'
statement, but this test case uses IMAGE_CLASSES+="testexport".

By the time the addition of testimage is evaluated, it is too late and
the inherit doesn't happen, so the test fails:

  Error for core-image-minimal.bb:do_testexport[depends], dependency
  ${TESTIMAGEDEPENDS} in '  cpio-native:do_populate_sysroot
  ${TESTIMAGEDEPENDS}' does not contain exactly one ':' character.

TESTIMAGEDEPENDS isn't set as testimage isn't inherited, so it doesn't
get expanded.

The documentation says to INHERIT += "testexport", and doing that means
the inherits work correctly.

Note that this works on the autobuilder because the local.conf always
has IMAGE_CLASSES += "testimage" added.

Signed-off-by: Ross Burton <ross.burton@arm.com>