bitbake: process: Improve client disconnect/idle sync
When clients are disconnecting, we need to ensure things are in sync with
the idle thread. Use a threading event to show when things drop back to
idle (there is only one always present idle handler for inotify). Later
patches should be able to clean up the inotify handler to work differently
and avoid hardcoded values.

If we're not idle at client disconnect, shut down the server to solve
the dilemma. Not ideal but shouldn't happen much in practise and would be
restarted.

(Bitbake rev: f719060d4834ca8391b5b979e5e32048c04d1107)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>