bitbake: fetch2/local: Mention the value of localpath in failure message
We have an obtuse rare and intermittent bitbake fetch failure. Understanding
the value of ud.localpath at the time of failure would be helpful to debug it
further. Tweak the debugging to provide this in the output.

(Bitbake rev: f1ffcfe0d90c700feae7a583b63ee02b91a97a54)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>