insane.bbclass: add a RECIPE_MAINTAINER check (oe-core recipes only)
Absent maintainer entries are as well a frequent source of friction, as they are checked
only in selftest, and so aren't revealed until autobuilder runs.

The selftest is retained as it also checks for obsolete entries in maintainers.inc
(not possible to do in insane class).

(From OE-Core rev: f79cedf360bbcb2ce2c358ee507091dccbc6c9f7)

Signed-off-by: Alexander Kanavin <alex@linutronix.de>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>