qemu: add PACKAGECONFIG for debuginfo support
qemu was auto-detecting the presence of libdw from elfutils, which
causes problems with a sstate object for qemu-native built with libdw is
ran on a machine without libdw.

We have no need for the debuginfo support to be enabled, so disable this
by default.

(From OE-Core rev: 3acd9fcd2bf4ad306455c275ec6ecbb4262fce22)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>